Refactor the commands for correct binary name (`pleasant`) - Add `quiet` flag to `ps create entry` command to minimise output so we can use it in scripts - Add `GetEntryIdByName` to `entryRepository`
