+++ cmd/pleasant/commands/create.go
@@ -1,5 +1,5 @@
-// Package pleasant /*
-package pleasant
+// Package commands /*
+package commands
 
 /*
 Copyright © 2021 David Lukac <david.lukac@users.noreply.github.com>
@@ -18,16 +18,12 @@ limitations under the License.
 */
 
 import (
-	"fmt"
-	"github.com/davidlukac/go-pleasant-cli/internal"
-	"strings"
-
 	"github.com/spf13/cobra"
 )
 
-// folderCmd represents the folder command
-var folderCmd = &cobra.Command{
-	Use:   "folder",
+// createCmd represents the create command
+var createCmd = &cobra.Command{
+	Use:   "create",
 	Short: "A brief description of your command",
 	Long: `A longer description that spans multiple lines and likely contains examples
 and usage of using your command. For example:
@@ -35,28 +31,8 @@ and usage of using your command. For example:
 Cobra is a CLI library for Go that empowers applications.
 This application is a tool to generate the needed files
 to quickly create a Cobra application.`,
-	//Args: func(cmd *cobra.Command, args []string) error {
-	//
-	//},
-	Args: cobra.ExactArgs(1),
-	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Printf("folder called with %s\n", args[0])
-		parts := strings.Split(args[0], "/")
-		exists := internal.FoldersExist(parts, "")
-		fmt.Println(exists)
-	},
 }
 
 func init() {
-	existsCmd.AddCommand(folderCmd)
-
-	// Here you will define your flags and configuration settings.
-
-	// Cobra supports Persistent Flags which will work for this command
-	// and all subcommands, e.g.:
-	// folderCmd.PersistentFlags().String("foo", "", "A help for foo")
-
-	// Cobra supports local flags which will only run when this command
-	// is called directly, e.g.:
-	// folderCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
+	passwordServerCmd.AddCommand(createCmd)
 }
