+++ cmd/pleasant/commands/exists_folder.go
@@ -1,5 +1,5 @@
-// Package pleasant /*
-package pleasant
+// Package commands /*
+package commands
 
 /*
 Copyright © 2021 David Lukac <david.lukac@users.noreply.github.com>
@@ -18,12 +18,16 @@ limitations under the License.
 */
 
 import (
+	"fmt"
+	"github.com/davidlukac/go-pleasant-cli/internal"
+	"strings"
+
 	"github.com/spf13/cobra"
 )
 
-// getCmd represents the get command
-var getCmd = &cobra.Command{
-	Use:   "get",
+// folderCmd represents the folder command
+var folderCmd = &cobra.Command{
+	Use:   "folder",
 	Short: "A brief description of your command",
 	Long: `A longer description that spans multiple lines and likely contains examples
 and usage of using your command. For example:
@@ -31,21 +35,28 @@ and usage of using your command. For example:
 Cobra is a CLI library for Go that empowers applications.
 This application is a tool to generate the needed files
 to quickly create a Cobra application.`,
-	//Run: func(cmd *cobra.Command, args []string) {
-	//	fmt.Printf("get called with %s\n", args)
+	//Args: func(cmd *cobra.Command, args []string) error {
+	//
 	//},
+	Args: cobra.ExactArgs(1),
+	Run: func(cmd *cobra.Command, args []string) {
+		fmt.Printf("folder called with %s\n", args[0])
+		parts := strings.Split(args[0], "/")
+		exists := internal.FoldersExist(parts, "")
+		fmt.Println(exists)
+	},
 }
 
 func init() {
-	passwordServerCmd.AddCommand(getCmd)
+	existsCmd.AddCommand(folderCmd)
 
 	// Here you will define your flags and configuration settings.
 
 	// Cobra supports Persistent Flags which will work for this command
 	// and all subcommands, e.g.:
-	// getCmd.PersistentFlags().String("foo", "", "A help for foo")
+	// folderCmd.PersistentFlags().String("foo", "", "A help for foo")
 
 	// Cobra supports local flags which will only run when this command
 	// is called directly, e.g.:
-	// getCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
+	// folderCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
 }
