+++ cmd/pleasant/commands/get_entry.go
@@ -1,5 +1,5 @@
-// Package pleasant /*
-package pleasant
+// Package commands /*
+package commands
 
 /*
 Copyright © 2021 David Lukac <david.lukac@users.noreply.github.com>
@@ -18,12 +18,15 @@ limitations under the License.
 */
 
 import (
+	"fmt"
+	"github.com/davidlukac/go-pleasant-cli/internal"
+	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 )
 
-// createCmd represents the create command
-var createCmd = &cobra.Command{
-	Use:   "create",
+// entryCmd represents the entry command
+var entryCmd = &cobra.Command{
+	Use:   "entry",
 	Short: "A brief description of your command",
 	Long: `A longer description that spans multiple lines and likely contains examples
 and usage of using your command. For example:
@@ -31,21 +34,23 @@ and usage of using your command. For example:
 Cobra is a CLI library for Go that empowers applications.
 This application is a tool to generate the needed files
 to quickly create a Cobra application.`,
-	//Run: func(cmd *cobra.Command, args []string) {
-	//	fmt.Println("create called")
-	//},
+	Args: cobra.ExactArgs(1),
+	Run: func(cmd *cobra.Command, args []string) {
+		log.Infoln(fmt.Sprintf("get entry called with %s", args[0]))
+		fmt.Println(internal.GetEntry(args[0]))
+	},
 }
 
 func init() {
-	passwordServerCmd.AddCommand(createCmd)
+	getCmd.AddCommand(entryCmd)
 
 	// Here you will define your flags and configuration settings.
 
 	// Cobra supports Persistent Flags which will work for this command
 	// and all subcommands, e.g.:
-	// createCmd.PersistentFlags().String("foo", "", "A help for foo")
+	// entryCmd.PersistentFlags().String("foo", "", "A help for foo")
 
 	// Cobra supports local flags which will only run when this command
 	// is called directly, e.g.:
-	// createCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
+	// entryCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
 }
